Refactor SessionArtifact tests
