+++ tests/unit/graph_reader/expected/complex_graph_h.py
@@ -1,0 +1,23 @@
+def get_h():
+    a0 = 0
+    a0 += 1
+    a = 1
+    a += 1
+    b = a * 2 + a0
+    c = b + 3
+    d = a * 4
+    e = d + 5
+    e += 6
+    a += 1
+    g = c + e * 2
+    h = a + g
+    return h
+
+
+def run_all():
+    h = get_h()
+    return h
+
+
+if __name__ == "__main__":
+    run_all()
