+++ tests/unit/graph_reader/inputs/complex.py
@@ -1,0 +1,28 @@
+import lineapy
+
+art = {}
+a0 = 0
+a0 += 1
+art["a0"] = lineapy.save(a0, "a0")
+a = 1
+art["a"] = lineapy.save(a, "a")
+
+a += 1
+b = a * 2 + a0
+c = b + 3
+d = a * 4
+e = d + 5
+e += 6
+art["c"] = lineapy.save(c, "c")
+art["e"] = lineapy.save(e, "e")
+
+f = c + 7
+art["f"] = lineapy.save(f, "f")
+a += 1
+g = c + e * 2
+art["g2"] = lineapy.save(g, "g2")
+h = a + g
+art["h"] = lineapy.save(h, "h")
+z = [1]
+z.append(h)
+art["z"] = lineapy.save(z, "z")
