+++ tests/unit/graph_reader/inputs/extract_common.py
@@ -1,0 +1,9 @@
+import lineapy
+
+art = {}
+a = 1
+a += 1
+b = a + 1
+art["b"] = lineapy.save(b, "b")
+c = a + 2
+art["c"] = lineapy.save(c, "c")
