+++ tests/unit/graph_reader/inputs/mutate_arter_save.py
@@ -1,0 +1,8 @@
+import lineapy
+
+art = {}
+a = [1]
+art["a"] = lineapy.save(a, "a")
+a.append(2)
+b = a[-1] + 1
+art["b"] = lineapy.save(b, "b")
